move Charger and ChargingInfrastructure from data to infrastructure
