+++ contribs/drt/src/main/java/org/matsim/contrib/edrt/schedule/EDrtChargingTask.java
@@ -20,9 +20,9 @@
 package org.matsim.contrib.edrt.schedule;
 
 import org.matsim.contrib.drt.schedule.DrtTask;
-import org.matsim.contrib.ev.data.Charger;
-import org.matsim.contrib.ev.fleet.ElectricVehicle;
 import org.matsim.contrib.ev.dvrp.ChargingTaskImpl;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
+import org.matsim.contrib.ev.infrastructure.Charger;
 
 public class EDrtChargingTask extends ChargingTaskImpl implements DrtTask {
 	public EDrtChargingTask(double beginTime, double endTime, Charger charger, ElectricVehicle ev, double totalEnergy) {
