+++ contribs/ev/src/main/java/org/matsim/contrib/ev/charging/ChargingEstimations.java
@@ -18,10 +18,10 @@
 
 package org.matsim.contrib.ev.charging;
 
-import org.matsim.contrib.ev.data.Charger;
-import org.matsim.contrib.ev.fleet.ElectricVehicle;
-
 import java.util.stream.Stream;
+
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
+import org.matsim.contrib.ev.infrastructure.Charger;
 
 /**
  * TODO Does not include future energy demand (e.g. AUX for plugged/queued vehs)
