+++ contribs/ev/src/main/java/org/matsim/contrib/ev/charging/ChargingStartEvent.java
@@ -19,12 +19,12 @@
 
 package org.matsim.contrib.ev.charging;
 
+import java.util.Map;
+
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.events.Event;
-import org.matsim.contrib.ev.data.Charger;
 import org.matsim.contrib.ev.fleet.ElectricVehicle;
-
-import java.util.Map;
+import org.matsim.contrib.ev.infrastructure.Charger;
 
 public class ChargingStartEvent extends Event {
 	public static final String EVENT_TYPE = "charging_start";
