+++ contribs/ev/src/main/java/org/matsim/contrib/ev/infrastructure/ChargingInfrastructure.java
@@ -1,8 +1,9 @@
-/* *********************************************************************** *
+/*
+ * *********************************************************************** *
  * project: org.matsim.*
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2018 by the members listed in the COPYING,        *
+ * copyright       : (C) 2019 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -14,22 +15,23 @@
  *   (at your option) any later version.                                   *
  *   See also COPYING, LICENSE and WARRANTY file                           *
  *                                                                         *
- * *********************************************************************** */
+ * *********************************************************************** *
+ */
 
-package org.matsim.contrib.ev.data;
+package org.matsim.contrib.ev.infrastructure;
+
+import java.util.Map;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.contrib.ev.charging.ChargingLogic;
 import org.matsim.core.api.experimental.events.EventsManager;
 
-import java.util.Map;
-
 /**
  * @author michalm
  */
 public interface ChargingInfrastructure {
-	public static final String CHARGERS = "chargers";
+	String CHARGERS = "chargers";
 
 	Map<Id<Charger>, Charger> getChargers();
