+++ contribs/taxi/src/main/java/org/matsim/contrib/etaxi/optimizer/BestChargerFinder.java
@@ -22,7 +22,7 @@ package org.matsim.contrib.etaxi.optimizer;
 import java.util.stream.Stream;
 
 import org.matsim.contrib.dvrp.fleet.DvrpVehicle;
-import org.matsim.contrib.ev.data.Charger;
+import org.matsim.contrib.ev.infrastructure.Charger;
 import org.matsim.contrib.taxi.optimizer.BestDispatchFinder;
 import org.matsim.contrib.taxi.optimizer.BestDispatchFinder.Dispatch;
 import org.matsim.contrib.util.LinkProvider;
